requirements.txt
asgiref==3.7.2
Django>=4.2.7
djangorestframework==3.14.0
django-filter==23.3
drf-yasg==1.21.7
sqlparse==0.4.4
psycopg2==2.9.9
